Update composite / object bvh divide
